+++ src/CoreShop/Component/Index/Interpreter/QuantityValueInterpreter.php
@@ -12,6 +12,7 @@
 
 namespace CoreShop\Component\Index\Interpreter;
 
+use CoreShop\Component\Index\Model\IndexableInterface;
 use CoreShop\Component\Index\Model\IndexColumnInterface;
 use Pimcore\Model\DataObject\Data\QuantityValue;
 
@@ -20,7 +21,7 @@ class QuantityValueInterpreter implements InterpreterInterface
     /**
      * {@inheritdoc}
      */
-    public function interpret($value, IndexColumnInterface $config = null)
+    public function interpret($value, IndexableInterface $object, IndexColumnInterface $config)
     {
         if ($value instanceof QuantityValue) {
             return $value->getValue();