+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Within V2
 
 ## 2.0.0-alpha.4 to 2.0.0-beta.1
+ - **BC break** Signature of following interfaces changed:
+    - ```CoreShop\Component\Index\Interpreter\InterpreterInterface```: public function interpret($value, IndexableInterface $object, IndexColumnInterface $config);
+    - ```CoreShop\Component\Index\Interpreter\LocalizedInterpreterInterface```: public function interpretForLanguage($language, $value, IndexableInterface $object, IndexColumnInterface $config);
+    - ```CoreShop\Component\Index\Interpreter\RelationInterpreterInterface```: public function interpretRelational($value, IndexableInterface $indexable, IndexColumnInterface $config);
+
  - **BC break** CoreShop now takes advantage of the dependent bundle feature introduced in Pimcore 5.1.2. Therefore,
  all bundles are now automatically loaded. This is a BC break, as when updating, you might run into issues.
  To solve the issues, check following things: